Gets a line and pushes it to execution helper

+++ get_line.c
@@ -12,12 +12,12 @@ int main(int argc, char *argv[], char *envp[])
 {
 	char *buffer;
 	size_t buffer_size = 0;
-	size_t chars_printed;
+	int chars_printed;
 
 	(void)argc;
 	(void)argv;
 
-	while (chars_printed != (size_t)EOF)
+	while (chars_printed != EOF)
 	{
 		write(1, "$ ", sizeof(char) * 2);
 		chars_printed = getline(&buffer, &buffer_size, stdin);
